HP-2541 created separate module for ConsumptionConfiguration for storing all related files
