+++ src/module/ConsumptionConfiguration/Domain/Collection/ConsumptionConfiguratorDataCollectionInterface.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace hipanel\modules\finance\helpers\ConsumptionConfiguration;
+namespace hipanel\modules\finance\module\ConsumptionConfiguration\Domain\Collection;
 
 interface ConsumptionConfiguratorDataCollectionInterface extends \IteratorAggregate
 {
