+++ src/module/ConsumptionConfiguration/Domain/Collection/ConsumptionConfiguratorDataCollectionTariffDecorator.php
@@ -1,7 +1,10 @@
 <?php declare(strict_types=1);
 
-namespace hipanel\modules\finance\helpers\ConsumptionConfiguration;
+namespace hipanel\modules\finance\module\ConsumptionConfiguration\Domain\Collection;
 
+use hipanel\modules\finance\module\ConsumptionConfiguration\Domain\ModelRegistry;
+use hipanel\modules\finance\module\ConsumptionConfiguration\Domain\Factory\ConsumptionConfiguratorDataFactory;
+use hipanel\modules\finance\module\ConsumptionConfiguration\Domain\TariffPriceTypeSpecification;
 use Traversable;
 
 /**
@@ -10,11 +13,13 @@ use Traversable;
  */
 class ConsumptionConfiguratorDataCollectionTariffDecorator implements ConsumptionConfiguratorDataCollectionInterface
 {
-    private TariffResourceHelper $helper;
+    private TariffPriceTypeSpecification $specification;
+    private ModelRegistry $modelRegistry;
 
     public function __construct(private readonly ConsumptionConfiguratorDataCollectionInterface $collection)
     {
-        $this->helper = new TariffResourceHelper();
+        $this->specification = new TariffPriceTypeSpecification();
+        $this->modelRegistry = new ModelRegistry();
     }
 
     public function getIterator(): Traversable
@@ -24,13 +29,13 @@ class ConsumptionConfiguratorDataCollectionTariffDecorator implements Consumptio
 
     private function addTariffToConfiguration(Traversable $configurations): array
     {
-        list ($defaultModel, $defaultResourceModel) = $this->helper->getDefaultModels();
+        list ($defaultModel, $defaultResourceModel) = $this->modelRegistry->getDefaultModels();
 
         $configurations = iterator_to_array($configurations);
 
         $configurations['tariff'] = ConsumptionConfiguratorDataFactory::create(
             'Tariff resources',
-            $this->helper->getTariffColumns(),
+            $this->specification->getTariffColumns(),
             [],
             $defaultModel,
             $defaultResourceModel,
