+++ src/module/ConsumptionConfiguration/Domain/Data/ConsumptionConfiguratorData.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace hipanel\modules\finance\helpers\ConsumptionConfiguration;
+namespace hipanel\modules\finance\module\ConsumptionConfiguration\Domain\Data;
 
 use hipanel\modules\client\models\Client;
 use hipanel\modules\finance\models\proxy\Resource as ResourceProxy;
